requirements.txt
# MLX Whisperer dependencies for Mac M3
mlx-whisper>=0.3.0
mlx>=0.5.0
numpy>=1.24.0

# API Server dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Audio processing
soundfile>=0.12.1
librosa>=0.10.1

# Utilities
pydantic>=2.5.0
python-dotenv>=1.0.0
pyyaml>=6.0.1
